Move data file wrangling code to CLI module
Split corresponding tests out into their own test module.
Since checking the input file name is something that is always done when
using the input_parser, make the find_input_file_name function call into
an argparse.Action.
