+++ src/tristan/command_line/images.py
@@ -19,12 +19,10 @@ from ..data import (
     cue_keys,
     cue_times,
     cues,
-    data_files,
     event_location_key,
     event_time_key,
     fem_falling,
     fem_rising,
-    find_file_names,
     first_cue_time,
     latrd_data,
     lvds_falling,
@@ -33,7 +31,14 @@ from ..data import (
     ttl_falling,
     ttl_rising,
 )
-from . import exposure_parser, image_output_parser, input_parser, version_parser
+from . import (
+    check_output_file,
+    data_files,
+    exposure_parser,
+    image_output_parser,
+    input_parser,
+    version_parser,
+)
 
 triggers = {
     "TTL-rising": ttl_rising,
@@ -75,14 +80,14 @@ def exposure(
 
 def single_image_cli(args):
     """Utility for making a single image from event-mode data."""
-    data_dir, root, output_file = find_file_names(
-        args.input_file, args.output_file, "single_image", args.force
+    output_file = check_output_file(
+        args.output_file, args.root, "single_image", args.force
     )
-    nexus_file = data_dir / f"{root}.nxs"
-    if nexus_file.exists():
+    input_nexus = args.data_dir / f"{args.root}.nxs"
+    if input_nexus.exists():
         # Write output NeXus file if we have an input NeXus file.
         output_nexus = CopyTristanNexus.single_image_nexus(
-            output_file, nexus_file, write_mode="w" if args.force else "x"
+            output_file, input_nexus, write_mode="w" if args.force else "x"
         )
     else:
         output_nexus = None
@@ -94,9 +99,9 @@ def single_image_cli(args):
     if args.image_size:
         image_size = tuple(map(int, args.image_size.split(",")))[::-1]
     else:
-        image_size = determine_image_size(nexus_file)
+        image_size = determine_image_size(input_nexus)
 
-    raw_files, _ = data_files(data_dir, root)
+    raw_files, _ = data_files(args.data_dir, args.root)
 
     keys = (event_location_key, event_time_key) + cue_keys
     with latrd_data(raw_files, keys=keys) as data:
@@ -123,11 +128,9 @@ def multiple_images_cli(args):
     The time between the start and end of the data collection is subdivided into a
     number of exposures of equal duration, providing a chronological stack of images.
     """
-    data_dir, root, output_file = find_file_names(
-        args.input_file, args.output_file, "images", args.force
-    )
-    nexus_file = data_dir / f"{root}.nxs"
-    if not nexus_file.exists():
+    output_file = check_output_file(args.output_file, args.root, "images", args.force)
+    input_nexus = args.data_dir / f"{args.root}.nxs"
+    if not input_nexus.exists():
         print(
             "Could not find a NeXus file containing experiment metadata.\n"
             "Resorting to writing raw image data without accompanying metadata."
@@ -136,9 +139,9 @@ def multiple_images_cli(args):
     if args.image_size:
         image_size = tuple(map(int, args.image_size.split(",")))[::-1]
     else:
-        image_size = determine_image_size(nexus_file)
+        image_size = determine_image_size(input_nexus)
 
-    raw_files, _ = data_files(data_dir, root)
+    raw_files, _ = data_files(args.data_dir, args.root)
 
     keys = (event_location_key, event_time_key) + cue_keys
     with latrd_data(raw_files, keys=keys) as data:
@@ -183,11 +186,11 @@ def multiple_images_cli(args):
             )
             images.store(data_set)
 
-    if nexus_file.exists():
+    if input_nexus.exists():
         # Write output NeXus file if we have an input NeXus file.
         output_nexus = CopyTristanNexus.multiple_images_nexus(
             output_file,
-            nexus_file,
+            input_nexus,
             nbins=num_images,
             write_mode="w" if args.force else "x",
         )
@@ -206,14 +209,12 @@ def pump_probe_cli(args):
     aggregated, providing a single stack of images that captures the evolution of the
     response of the measurement to a pump signal.
     """
-    data_dir, root, output_file = find_file_names(
-        args.input_file, args.output_file, "images", args.force
-    )
-    nexus_file = data_dir / f"{root}.nxs"
-    if nexus_file.exists():
+    output_file = check_output_file(args.output_file, args.root, "images", args.force)
+    input_nexus = args.data_dir / f"{args.root}.nxs"
+    if input_nexus.exists():
         # Write output NeXus file if we have an input NeXus file.
         output_nexus = CopyTristanNexus.pump_probe_nexus(
-            output_file, nexus_file, write_mode="w" if args.force else "x"
+            output_file, input_nexus, write_mode="w" if args.force else "x"
         )
     else:
         output_nexus = None
@@ -225,9 +226,9 @@ def pump_probe_cli(args):
     if args.image_size:
         image_size = tuple(map(int, args.image_size.split(",")))[::-1]
     else:
-        image_size = determine_image_size(nexus_file)
+        image_size = determine_image_size(input_nexus)
 
-    raw_files, _ = data_files(data_dir, root)
+    raw_files, _ = data_files(args.data_dir, args.root)
 
     keys = (event_location_key, event_time_key) + cue_keys
     with latrd_data(raw_files, keys=keys) as data:
